+++ docs/pages/components/autocomplete/Autocomplete.vue
@@ -11,6 +11,14 @@
 
         <Example :component="ExObjArray" :code="ExObjArrayCode" title="Object array"/>
 
+        <Example :component="ExHeader" :code="ExHeaderCode" title="Header">
+            <div class="tags has-addons">
+                <span class="tag is-success">New!</span>
+                <span class="tag is-info">0.6.6</span>
+            </div>
+            <p>You can add your custom header to the autocomplete.</p>
+        </Example>
+
         <Example :component="ExCustomAsync" :code="ExCustomAsyncCode" title="Async with custom template">
             <p>You can have a custom template by adding a scoped slot to it.</p>
             <p><small>API from <a href="https://www.themoviedb.org" target="_blank">TMDb</a></small>.</p>
@@ -29,6 +37,9 @@
     import ExObjArray from './examples/ExObjArray'
     import ExObjArrayCode from '!!raw-loader!./examples/ExObjArray'
 
+    import ExHeader from './examples/ExHeader'
+    import ExHeaderCode from '!!raw-loader!./examples/ExHeader'
+
     import ExCustomAsync from './examples/ExCustomAsync'
     import ExCustomAsyncCode from '!!raw-loader!./examples/ExCustomAsync'
 
@@ -40,6 +51,8 @@
                 ExSimpleCode,
                 ExObjArrayCode,
                 ExObjArray,
+                ExHeader,
+                ExHeaderCode,
                 ExCustomAsyncCode,
                 ExCustomAsync
             }